constexpr, Python stable API, header-only, nits. (#11)
This PR adds the below:
- The python module is now PEP-384 compliant and uses the Python stable ABI which means that it should be compatible with every current and future Python 3 versions (#12),
"""
Currently, each feature release introduces a new name for the Python DLL on Windows, and may cause incompatibilities for extension modules on Unix. This PEP proposes to define a stable set of API functions which are guaranteed to be available for the lifetime of Python 3, and which will also remain binary-compatible across versions. Extension modules and applications embedding Python can work with different feature releases as long as they restrict themselves to this stable ABI.
"""
- Turned the library into a header-only library (#13),
- `constexpr` the library wherever was possible, as well as `const`,
- Fix a bug where the `mmap` return was checked but it didn't invoke `perror`,
- Fix the `DisplayUtils` mess,
- Turn `Page` into a namespace instead of a structure,
- Use `optional` for `VirtTranslate` to handle error case properly,
- Build w/ Py3.6-3.9 on both Windows/Linux in the CI,
- Fix a bunch of small things